requirements.txt
# PCB缺陷检测项目依赖

# 深度学习框架
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# 模型量化
bitsandbytes>=0.41.0
autoawq>=0.1.8

# 数据处理
datasets>=2.14.0
pillow>=10.0.0
numpy>=1.24.0

# LoRA
peft>=0.6.0

# 评估指标
scikit-learn>=1.3.0

# API服务
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain (可选，用于智能体框架)
langchain>=0.1.0
langchain-community>=0.0.10

# LangGraph (可选，用于工作流)
langgraph>=0.0.40

# 向量数据库
chromadb>=0.4.15
sentence-transformers>=2.2.2

# 工具
pydantic>=2.0.0
requests>=2.31.0

# Qwen3-VL模型相关
qwen-vl-utils>=0.0.1

